+++ src/libraries/System.Threading.RateLimiting/src/System/Threading/RateLimiting/PartitionedRateLimiter.T.cs
@@ -127,6 +127,7 @@ namespace System.Threading.RateLimiting
         /// <typeparam name="TOuter">The type to translate into <typeparamref name="TResource"/>.</typeparam>
         /// <param name="keyAdapter">The function to be called every time a <typeparamref name="TOuter"/> is passed to
         /// PartitionedRateLimiter&lt;TOuter&gt;.Acquire(TOuter, int) or PartitionedRateLimiter&lt;TOuter&gt;.WaitAsync(TOuter, int, CancellationToken).</param>
+        /// <remarks><see cref="PartitionedRateLimiter{TResource}.Dispose()"/> or <see cref="PartitionedRateLimiter{TResource}.DisposeAsync()"/> does not dispose the wrapped <see cref="PartitionedRateLimiter{TResource}"/>.</remarks>
         /// <returns>A new PartitionedRateLimiter&lt;TOuter&gt; that translates <typeparamref name="TOuter"/>
         /// to <typeparamref name="TResource"/> and calls the inner <see cref="PartitionedRateLimiter{TResource}"/>.</returns>
         public PartitionedRateLimiter<TOuter> TranslateKey<TOuter>(Func<TOuter, TResource> keyAdapter)
@@ -135,61 +136,6 @@ namespace System.Threading.RateLimiting
             // Should the default be to dispose the inner limiter and have an option to not dispose it?
             // See Stream wrappers like SslStream for prior-art
             return new TranslatingLimiter<TResource, TOuter>(this, keyAdapter);
-        }
-    }
-
-    internal sealed class TranslatingLimiter<TInner, TResource> : PartitionedRateLimiter<TResource>
-    {
-        private readonly PartitionedRateLimiter<TInner> _innerRateLimiter;
-        private readonly Func<TResource, TInner> _keyAdapter;
-
-        private bool _disposed;
-
-        public TranslatingLimiter(PartitionedRateLimiter<TInner> inner, Func<TResource, TInner> keyAdapter)
-        {
-            _innerRateLimiter = inner;
-            _keyAdapter = keyAdapter;
-        }
-
-        public override int GetAvailablePermits(TResource resourceID)
-        {
-            ThrowIfDispose();
-            TInner key = _keyAdapter(resourceID);
-            return _innerRateLimiter.GetAvailablePermits(key);
-        }
-
-        protected override RateLimitLease AcquireCore(TResource resourceID, int permitCount)
-        {
-            ThrowIfDispose();
-            TInner key = _keyAdapter(resourceID);
-            return _innerRateLimiter.Acquire(key, permitCount);
-        }
-
-        protected override ValueTask<RateLimitLease> WaitAsyncCore(TResource resourceID, int permitCount, CancellationToken cancellationToken)
-        {
-            ThrowIfDispose();
-            TInner key = _keyAdapter(resourceID);
-            return _innerRateLimiter.WaitAsync(key, permitCount, cancellationToken);
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            _disposed = true;
-            base.Dispose(disposing);
-        }
-
-        protected override ValueTask DisposeAsyncCore()
-        {
-            _disposed = true;
-            return base.DisposeAsyncCore();
-        }
-
-        private void ThrowIfDispose()
-        {
-            if (_disposed)
-            {
-                throw new ObjectDisposedException(nameof(PartitionedRateLimiter));
-            }
         }
     }
 }
